Move codeFontFamily and remove fluent-theme usage
