+++ packages/tsx-editor/package.json
@@ -45,7 +45,6 @@
   },
   "dependencies": {
     "@microsoft/load-themed-styles": "^1.7.13",
-    "@uifabric/fluent-theme": "^7.1.4",
     "@uifabric/monaco-editor": "^0.2.0",
     "office-ui-fabric-react": "^7.36.0",
     "react-syntax-highlighter": "^10.1.3",
