+++ packages/tsx-editor/src/components/common.ts
@@ -1,0 +1,2 @@
+/** Font family for code snippets/editors */
+export const codeFontFamily = 'Monaco, Menlo, Consolas, "Droid Sans Mono", "Inconsolata", "Courier New", monospace';
