+++ packages/tsx-editor/src/components/index.ts
@@ -2,6 +2,6 @@ export { Editor } from './Editor';
 export * from './Editor.types';
 export { EditorWrapper } from './EditorWrapper';
 export * from './EditorWrapper.types';
-export { ITypeScriptSnippetProps, TypeScriptSnippet, codeFontFamily } from './TypeScriptSnippet';
+export { codeFontFamily } from './common';
 
 // Intentionally not exporting EditorError and EditorPreview since they're basically internal helpers
